# packages/types/src/generated/changeRoleIn.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HomeOps API
 * HomeOps backend API — household management SaaS (payments, documents, obligations).
 * OpenAPI spec version: 0.1.0
 */
import type { ChangeRoleInRole } from "./changeRoleInRole";

export interface ChangeRoleIn {
  role: ChangeRoleInRole;
}