# packages/types/src/generated/switchOut.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HomeOps API
 * HomeOps backend API — household management SaaS (payments, documents, obligations).
 * OpenAPI spec version: 0.1.0
 */
import type { HouseholdOut } from "./householdOut";

export interface SwitchOut {
  access_token?: string;
  household?: HouseholdOut;
  token_type?: string;
}